Make AJobScheduler.RunsEveryJobOnceEvenRunningOftenOnEmptyQueue pass
Multiple things happened here. This fixes the crash that happens when
you want to pop on an empty queue (getting the empty() and then pop() in
a concurrent environment is a recipie for a disaster).

I made it work with std::optional that is C++17.

Last but not least the 1000 jobs resulted a 63 sec test run on my
laptop. I reduced the number of jobs to 150 that is about 10 sec.
Probably you could also play with the sleep, but even this change made
the software crash on the previous implementation e.g. it is good enough
for a test.